+++ server/index.js
@@ -2,7 +2,7 @@ import dotenv from "dotenv";
 dotenv.config();
 import mongoose from "mongoose";
 import createApp from "./app.js";
-import connectDB from "./utils/connectDB.js";
+import connectDB from "./src/utils/connectDB.js";
 
 const app = createApp();
 connectDB(process.env.DATABASE_CONNECTION_STRING);
