+++ server/src/controllers/Opening.js
