+++ server/src/models/Opening.js
