+++ server/src/routes/Opening.js
