+++ server/src/routes/User.js
@@ -1,5 +1,5 @@
 import express from "express";
-import { createUser, updateResume } from "./../controllers/User.js";
+import { createUser, updateResume } from "../controllers/User.js";
 import { checkUserId } from "../middleware/User.js";
 const UserRouter = express.Router();
