+++ server/test/Opening/getAllOpening.js
@@ -1,8 +1,8 @@
 import request from "supertest";
 import createApp from "../../app";
 import mongoose from "mongoose";
-import Opening from "./../../models/Opening";
-import User from "./../../models/User";
+import Opening from "./../../src/models/Opening";
+import User from "./../../src/models/User";
 
 export default getAllOpeningTest = () => {
   describe("Get All Opening", () => {
